--- src/main/java/commands/Command.java
@@ -1,8 +1,0 @@
-package commands;
-
-import data.MyData;
-import ui.Ui;
-
-public abstract class Command {
-    public abstract void execute(MyData data, Ui ui);
-}
