--- src/main/java/commands/List.java
@@ -1,10 +1,0 @@
-package commands;
-
-import data.MyData;
-import ui.Ui;
-
-public class List extends Command {
-    public void execute(MyData data, Ui ui) {
-        ui.list(data);
-    }
-}
