+++ src/main/java/duke/commands/Command.java
@@ -1,0 +1,8 @@
+package duke.commands;
+
+import duke.data.MyData;
+import duke.ui.Ui;
+
+public abstract class Command {
+    public abstract void execute(MyData data, Ui ui);
+}
