+++ src/main/java/duke/commands/ListDate.java
@@ -1,10 +1,7 @@
-package commands;
+package duke.commands;
 
-import data.MyData;
-import tasks.Deadline;
-import tasks.Task;
-import tasks.Event;
-import ui.Ui;
+import duke.data.MyData;
+import duke.ui.Ui;
 
 import java.time.LocalDate;
