+++ src/main/java/duke/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package tasks;
+package duke.tasks;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
