+++ src/main/java/duke/tasks/Task.java
@@ -1,4 +1,4 @@
-package tasks;
+package duke.tasks;
 
 import java.time.format.DateTimeFormatter;
