+++ src/main/java/duke/tasks/ToDo.java
@@ -1,0 +1,17 @@
+package duke.tasks;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String toSave() {
+        return "T /" + super.toSave();
+    }
+}
