--- src/main/java/tasks/ToDo.java
@@ -1,17 +1,0 @@
-package tasks;
-
-public class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String toSave() {
-        return "T /" + super.toSave();
-    }
-}
